requirements.txt
numpy
opencv-python
pyserial
tessdata.fast-eng
tesserocr
